+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -334,7 +334,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KW9TA5DJTKJMN2JZ41DTNHXQ"
+        "01KWA3J1T54WD3RGATW0BSQ2FW"
     )
     r"""The unique identifier of the data source"""