+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsefunction.py
@@ -652,9 +652,9 @@ class EvaluatorResponseFunction(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-06-29T13:49:50.653Z"
+    created: Optional[str] = "2026-06-29T16:31:24.066Z"
 
-    updated: Optional[str] = "2026-06-29T13:49:50.653Z"
+    updated: Optional[str] = "2026-06-29T16:31:24.066Z"
 
     updated_by_id: OptionalNullable[str] = UNSET