+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsellm.py
@@ -388,9 +388,9 @@ class EvaluatorResponseLlm(BaseModel):
 
     mode: Mode
 
-    created: Optional[str] = "2026-06-29T13:49:50.653Z"
+    created: Optional[str] = "2026-06-29T16:31:24.066Z"
 
-    updated: Optional[str] = "2026-06-29T13:49:50.653Z"
+    updated: Optional[str] = "2026-06-29T16:31:24.066Z"
 
     updated_by_id: OptionalNullable[str] = UNSET