+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KW9TA5DF2ETP8H7WNRMM9BG9"
+        "01KWA3J1T095B4YCPGX11749QV"
     )
     r"""The unique identifier of the data source"""