+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasetop.py
@@ -160,7 +160,7 @@ class UpdateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-06-29T13:49:49.299Z")
+    updated: Optional[datetime] = parse_datetime("2026-06-29T16:31:22.968Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")