+++ packages/orq-rc/RELEASES.md
@@ -9059,3 +9059,13 @@ Based on:
 - [python v4.11.0-rc.55] packages/orq-rc
 ### Releases
 - [PyPI v4.11.0-rc.55] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.55 - packages/orq-rc
+
+## 2026-06-29 16:34:03
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.786.1 (2.913.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.11.0-rc.56] packages/orq-rc
+### Releases
+- [PyPI v4.11.0-rc.56] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.56 - packages/orq-rc